+++ lib/ahoy_email/processor.rb
@@ -2,14 +2,14 @@ module AhoyEmail
   class Processor
     attr_reader :mailer, :options
 
-    UTM_PARAMETERS = %w(utm_source utm_medium utm_term utm_content utm_campaign)
+    UTM_PARAMETERS = %w[utm_source utm_medium utm_term utm_content utm_campaign].freeze
 
     def initialize(mailer, options)
       @mailer = mailer
       @options = options
 
       unknown_keywords = options.keys - AhoyEmail.default_options.keys
-      raise ArgumentError, "unknown keywords: #{unknown_keywords.join(", ")}" if unknown_keywords.any?
+      raise ArgumentError, "unknown keywords: #{unknown_keywords.join(', ')}" if unknown_keywords.any?
     end
 
     def perform
@@ -26,7 +26,7 @@ module AhoyEmail
     end
 
     def token
-      @token ||= SecureRandom.urlsafe_base64(32).gsub(/[\-_]/, "").first(32)
+      @token ||= SecureRandom.urlsafe_base64(32).gsub(/[\-_]/, '').first(32).downcase
     end
 
     def track_message
@@ -44,9 +44,7 @@ module AhoyEmail
         data[:user_id] = id.is_a?(Integer) ? id : id.to_s
       end
 
-      if options[:open] || options[:click]
-        data[:token] = token
-      end
+      data[:token] = token if options[:open] || options[:click]
 
       if options[:utm_params]
         UTM_PARAMETERS.map(&:to_sym).each do |k|
@@ -63,12 +61,12 @@ module AhoyEmail
         regex = /<\/body>/i
         url =
           url_for(
-            controller: "ahoy/messages",
-            action: "open",
+            controller: 'ahoy/messages',
+            action: 'open',
             id: token,
-            format: "gif"
+            format: 'gif'
           )
-        pixel = ActionController::Base.helpers.image_tag(url, size: "1x1", alt: "")
+        pixel = ActionController::Base.helpers.image_tag(url, size: '1x1', alt: '')
 
         # try to add before body tag
         if raw_source.match(regex)
@@ -90,15 +88,16 @@ module AhoyEmail
           v: 1,
           tid: options[:google_analytics_code],
           cid: 555,
-          t: :event,
-          ec: :email,
-          ea: :open,
-          el: epath,
+          t: :pageview,
+          # t: :event,
+          # ec: :email,
+          # ea: :open,
+          # el: epath,
           dp: epath
         }
-        
+
         url = "https://www.google-analytics.com/collect?#{tracker.to_query}"
-        pixel = ActionController::Base.helpers.image_tag(url, size: "1x1", alt: "")
+        pixel = ActionController::Base.helpers.image_tag(url, size: '1x1', alt: '')
 
         # try to add before body tag
         if raw_source.match(regex)
@@ -114,32 +113,34 @@ module AhoyEmail
         body = (message.html_part || message).body
 
         doc = Nokogiri::HTML(body.raw_source)
-        doc.css("a[href]").each do |link|
-          uri = parse_uri(link["href"])
+        doc.css('a[href]').each do |link|
+          uri = parse_uri(link['href'])
           next unless trackable?(uri)
+
           # utm params first
-          if options[:utm_params] && !skip_attribute?(link, "utm-params")
+          if options[:utm_params] && !skip_attribute?(link, 'utm-params')
             params = uri.query_values(Array) || []
             UTM_PARAMETERS.each do |key|
               next if params.any? { |k, _v| k == key } || !options[key.to_sym]
+
               params << [key, options[key.to_sym]]
             end
             uri.query_values = params
-            link["href"] = uri.to_s
+            link['href'] = uri.to_s
           end
 
-          if options[:click] && !skip_attribute?(link, "click")
-            # TODO sign more than just url and transition to HMAC-SHA256
-            signature = OpenSSL::HMAC.hexdigest("SHA1", AhoyEmail.secret_token, link["href"])
-            link["href"] =
-              url_for(
-                controller: "ahoy/messages",
-                action: "click",
-                id: token,
-                url: link["href"],
-                signature: signature
-              )
-          end
+          next unless options[:click] && !skip_attribute?(link, 'click')
+
+          # TODO: sign more than just url and transition to HMAC-SHA256
+          signature = OpenSSL::HMAC.hexdigest('SHA1', AhoyEmail.secret_token, link['href'])
+          link['href'] =
+            url_for(
+              controller: 'ahoy/messages',
+              action: 'click',
+              id: token,
+              url: link['href'],
+              signature: signature
+            )
         end
 
         # hacky
@@ -157,7 +158,7 @@ module AhoyEmail
         # remove it
         link.remove_attribute(attribute)
         true
-      elsif link["href"].to_s =~ /unsubscribe/i && !options[:unsubscribe_links]
+      elsif link['href'].to_s =~ /unsubscribe/i && !options[:unsubscribe_links]
         # try to avoid unsubscribe links
         true
       else
@@ -167,14 +168,17 @@ module AhoyEmail
 
     # Filter trackable URIs, i.e. absolute one with http
     def trackable?(uri)
-      uri && uri.absolute? && %w(http https).include?(uri.scheme)
+      uri && uri.absolute? && %w[http https].include?(uri.scheme)
     end
 
     # Parse href attribute
     # Return uri if valid, nil otherwise
     def parse_uri(href)
       # to_s prevent to return nil from this method
-      Addressable::URI.heuristic_parse(href.to_s) rescue nil
+
+      Addressable::URI.heuristic_parse(href.to_s)
+    rescue StandardError
+      nil
     end
 
     def url_for(opt)